Refactored code, started on op mode incorporating CNNs
Moved the code to load and execute neural networks into a separate caffe-wrapper library
Refactored ftc-cnn-test to use the new separate library
Started making a CnnOpMode which will eventually use CNNs to perform tasks in autonomous mode.
