+++ caffe-wrapper/build.gradle
@@ -1,15 +1,15 @@
-apply plugin: 'com.android.application'
+apply plugin: 'com.android.library'
 
 android {
     compileSdkVersion 23
     buildToolsVersion "25.0.0"
 
     defaultConfig {
-        applicationId "io.github.codecube.cnntest"
         minSdkVersion 16
         targetSdkVersion 16
+        versionCode 1
+        versionName "1.0"
     }
-
     buildTypes {
         release {
             minifyEnabled false
@@ -27,5 +27,4 @@ repositories {
 dependencies {
     compile project(':opencv-java')
     compile project(':ftc-visionlib')
-    compile project(path: ':caffe-wrapper')
 }
