+++ caffe-wrapper/src/main/java/com/sh1r0/caffe_android_lib/CaffeMobile.java
