+++ caffe-wrapper/src/main/java/io/github/codecube/caffewrapper/CaffeCNN.java
@@ -1,4 +1,4 @@
-package io.github.codecube.cnntest;
+package io.github.codecube.caffewrapper;
 
 import android.os.Environment;
