+++ settings.gradle
@@ -1,4 +1,4 @@
-include ':FtcRobotController'
+include ':FtcRobotController', ':caffe-wrapper'
 include ':TeamCode'
 include ':opencv-java'
 include ':ftc-visionlib'
